Create util.py file
